Init Dashboard
This is the initial layout for the dashboard, some tests included for the LeadGrid component that is used. Also, dependencies and dev dependencies are updated, along with multiple config file changes or additions that are required for these tests to work, my reference was the Mantine Next.js App Routing Template that is based on using TypeScript, linked here: https://github.com/mantinedev/next-app-template/tree/master.