Dashboard Update
Adds buttons/links to admin page and fuel input form. At this point not positive if we should have the Login on the dashboard, since I think it makes things a little confusing as to whether or not you can access the other things on the dashboard depending on whether you're signed in or not. Also consolidated components that are definitely just for the dashboard in a folder. Linking works fine, also added some notes as to what next steps may be for the login and fuel input form considering the hydration and async errors we're getting with them.