Database Design
Attempting to parse together what the database schema/design is given the query calls given in Justin's ColdFusion code. Notes/documentation given, and a rough guess as to the database design given as a script. More notes given there, but I think that should be good enough for a mock database.