FuelForm Archival - Beginning of Auth
Changing this branch's (input-form) version of fuel input form to be a backup so it doesn't cause any conflicts with Michael's site-page-component branch later on, as we're going with his implementation.

Also gave rough break down of authorization steps, still in research phase to see if this is the right path for us.