Auth w/ Prisma DB Works
The authentication is setup, but right now it's a little too good at keeping you logged in... Restarting the server keeps you logged in, so does refreshing the page, to the point where the main page is basically inaccessible once you login once. This will be my next challenge, as I'll need to find a way to let users logout so they're not perpetually stuck logged in for one fuel site on their device/kiosk. But the auth works!