Init Serwist
Next-pwa was outdated and not working. Changed to serwist which is similar but more up to date. Still having trouble with it, setup looks fine but manifest.json is not being parsed correctly and I can't seem to fix that no matter how many times i check for errors or replace it completely. Guides say that there should be a prompt immediately when you open the program with the current setup, not sure if this is my machine having an issue, or what, but as far as I can tell this is the complete setup for the PWA.